requirements.txt
flask==3.1.0
gunicorn==23.0.0
pikepdf==10.3.0
Pillow==11.1.0
